compose a transport with an overriden roundtrip
